requirements.txt
discord-py-interactions
python-dotenv
psycopg2